requirements.txt
streamlit
pandas
plotly
requests
groq
cryptography
fpdf2